requirements.txt
accelerate==1.11.0
evaluate==0.4.6
faiss_cpu==1.12.0
fastapi==0.121.1
langchain==1.0.5
langchain_community==0.4.1
langchain_core==1.0.4
line_profiler_pycharm==1.2.0
loguru==0.7.3
neo4j==5.28.2
numpy==2.3.4
openai==2.7.2
PyYAML==6.0.2
PyYAML==6.0.3
redis==7.0.1
Requests==2.32.5
scikit_learn==1.7.2
scipy==1.16.3
tabulate==0.9.0
tiktoken==0.11.0
torch==2.8.0+cu128
tqdm==4.67.1
transformers==4.56.2
ujson==5.11.0
uvicorn==0.38.0
